+++ packages/visual-grid-client/test/util/FakeEyesWrapper.js
@@ -157,9 +157,9 @@ class FakeEyesWrapper extends EventEmitter {
         userAgent: browserName,
         deviceSize: deviceName && devices[deviceName],
         selectorRegions: selectorsToFindRegionsFor
-          ? selectorsToFindRegionsFor.map(
-              selector => selectorsToLocations[selector.selector || selector],
-            )
+          ? selectorsToFindRegionsFor.map(selector => [
+              selectorsToLocations[selector.selector || selector],
+            ])
           : undefined,
       })
     })