+++ packages/eyes-protractor/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 ## Unreleased
 
+- deprecate `saveDebugData`
 
+## 1.9.1 - 2020/10/15
+
+- add iosVersion to readme
+- updated to @applitools/eyes-sdk-core@12.4.2 (from 12.3.0)
+- updated to @applitools/visual-grid-client@15.0.10 (from 15.0.9)
 
 ## 1.9.0 - 2020/10/6