+++ packages/eyes-webdriverio-5-service/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- updated to @applitools/eyes-webdriverio@5.25.1 (from 5.24.0)
 
 
 ## 1.6.11 - 2020/10/7