+++ packages/sdk-shared/CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
+
+## 0.3.0 - 2020/10/14
+
 - add native mobile app device configurations for iOS and Android
 - add custom coverage tests to check native app viewport screenshots
+- updated to @applitools/sdk-coverage-tests@1.0.16 (from 1.0.14)
 
 ## 0.2.0 - 2020/10/7